// server/src/desposition/dto/desposition.interface.ts
export interface GetDespositionInterface {
    MCD_DYSPOZYCJE_ID: number
    MAGAZYN_DST_ID: number
    WYDZIALY_ID: number
    NUMER_DYSPOZYCJI: string
    SERIA_ID: number
    LIMIT_ID: number
    PZP_ID: number
    STATUS: number
    INS_DATE: string
    MAGAZYNY_KOD: string
    WYDZIALY_KOD: string
    SERIA_KOD: string
    LIMIT_KOD: string
    PZP_KOD: string
}